--- apps/data/src/pipelines/centrum_pipeline.py
@@ -1,1540 +1,0 @@
-# Databricks notebook source
-# DBTITLE 1,openJII Medallion Architecture Pipeline
-# This notebook implements the complete medallion architecture (Bronze-Silver-Gold)
-# for openJII IoT sensor data processing following the dual medallion pattern
-
-# COMMAND ----------
-import dlt
-from pyspark.sql import functions as F
-from pyspark.sql.types import StructType, StructField, StringType, DoubleType, TimestampType, ArrayType, LongType
-import requests
-import pandas as pd
-from datetime import datetime
-from enrich.user_metadata import add_user_column
-from enrich.device_metadata import add_device_registry
-from enrich.annotations_metadata import add_annotation_column
-from enrich.custom_metadata import add_custom_metadata_column
-from enrich.macro_execution import make_execute_macro_udf
-from openjii import decompress_sample
-from data_repair import apply_inline_repairs
-
-# COMMAND ----------
-
-# DBTITLE 1,Schema Definition
-# Define schema for sensor data to parse JSON payloads
-question_schema = StructType([
-    StructField("question_label", StringType(), True),
-    StructField("question_text", StringType(), True),
-    StructField("question_answer", StringType(), True)
-])
-
-macro_schema = StructType([
-    StructField("id", StringType(), True),
-    StructField("name", StringType(), True),
-    StructField("filename", StringType(), True)
-])
-
-# Define annotation schema to match the database structure
-annotation_content_schema = StructType([
-    StructField("text", StringType(), True),
-    StructField("flagType", StringType(), True)
-])
-
-annotation_schema = StructType([
-    StructField("id", StringType(), True),
-    StructField("rowId", StringType(), True),
-    StructField("type", StringType(), True),
-    StructField("content", annotation_content_schema, True),
-    StructField("createdBy", StringType(), True),
-    StructField("createdByName", StringType(), True),
-    StructField("createdAt", TimestampType(), True),
-    StructField("updatedAt", TimestampType(), True)
-])
-
-sensor_schema = StructType([
-    StructField("topic", StringType(), False),
-    StructField("device_name", StringType(), True),
-    StructField("device_version", StringType(), True),
-    StructField("device_id", StringType(), True),
-    StructField("device_battery", DoubleType(), True),
-    StructField("device_firmware", StringType(), True),
-    StructField("sample", StringType(), True),
-    StructField("_sample_encoding", StringType(), True),
-    StructField("timestamp", TimestampType(), False),
-    StructField("output", StringType(), True),
-    StructField("questions", ArrayType(question_schema), True),
-    StructField("user_id", StringType(), True),
-    StructField("timezone", StringType(), True),
-    StructField("macros", ArrayType(macro_schema), True),
-    StructField("annotations", ArrayType(annotation_schema), True),
-    # One uuid per multi-device workbook run; the round's rows share it.
-    # Nullable: absent on single-device uploads and all older payloads.
-    StructField("workbook_run_id", StringType(), True)
-])
-
-# COMMAND ----------
-
-# DBTITLE 1,Configuration
-
-ENVIRONMENT = spark.conf.get("ENVIRONMENT", "dev").lower()
-CATALOG_NAME = spark.conf.get("CATALOG_NAME")
-
-BRONZE_TABLE = spark.conf.get("BRONZE_TABLE", "raw_data")
-SILVER_TABLE = spark.conf.get("SILVER_TABLE", "clean_data")
-
-KINESIS_STREAM_NAME = spark.conf.get("KINESIS_STREAM_NAME")
-CHECKPOINT_PATH = spark.conf.get("CHECKPOINT_PATH")
-SERVICE_CREDENTIAL_NAME = spark.conf.get("SERVICE_CREDENTIAL_NAME")
-MONITORING_SLACK_CHANNEL = spark.conf.get("MONITORING_SLACK_CHANNEL")
-
-EXPERIMENT_STATUS_TABLE = "experiment_status"
-EXPERIMENT_RAW_DATA_TABLE = "experiment_raw_data"
-EXPERIMENT_DEVICE_DATA_TABLE = "experiment_device_data"
-EXPERIMENT_MACRO_DATA_TABLE = "experiment_macro_data"
-EXPERIMENT_CONTRIBUTORS_TABLE = "experiment_contributors"
-EXPERIMENT_DEVICES_TABLE = "experiment_devices"
-EXPERIMENT_TABLE_METADATA = "experiment_table_metadata"
-ENRICHED_RAW_DATA_VIEW = "enriched_experiment_raw_data"
-ENRICHED_MACRO_DATA_VIEW = "enriched_experiment_macro_data"
-RAW_IMPORTED_DATA_TABLE = "raw_imported_data"
-RAW_UPLOADED_DATA_TABLE = "raw_uploaded_data"
-EXPERIMENT_UPLOADED_DATA_TABLE = "experiment_uploaded_data"
-ENRICHED_UPLOADED_DATA_VIEW = "enriched_experiment_uploaded_data"
-ANNOTATIONS_SOURCE_TABLE = "experiment_annotations_source"
-METADATA_SOURCE_TABLE = "experiment_metadata_source"
-
-# Legacy macro identifiers that need mapping to actual macro UUIDs.
-# Some data was sent prior to the implementation of proper macro ID handling, so we maintain a mapping of legacy identifiers to actual UUIDs for correct processing.
-LEGACY_MACRO_ID_MAP_PATH = f"/Volumes/{CATALOG_NAME}/centrum/data-legacy/internal/legacy_macro_id_map.json"
-
-def _load_legacy_macro_id_map() -> dict[str, str]:
-    """Load legacy→UUID mapping from a volume JSON file. Returns empty dict on failure."""
-    import json
-    try:
-        with open(LEGACY_MACRO_ID_MAP_PATH) as f:
-            mapping = json.load(f)
-        print(f"[INFO] Loaded {len(mapping)} legacy macro ID mappings from {LEGACY_MACRO_ID_MAP_PATH}")
-        return mapping
-    except Exception as e:
-        print(f"[WARN] Legacy macro ID map not found or unreadable at {LEGACY_MACRO_ID_MAP_PATH}: {e}")
-        return {}
-
-LEGACY_MACRO_ID_MAP: dict[str, str] = _load_legacy_macro_id_map()
-
-# Backend's macro batch endpoint requires UUID macro_ids; non-UUIDs trigger
-# request-wide 400s. Gate at gold so bad rows never reach the UDF.
-MACRO_ID_UUID_PATTERN = r"(?i)\A[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}\z"
-
-# COMMAND ----------
-
-# DBTITLE 1,Bronze Layer - Raw Data Processing
-@dlt.table(
-    name=BRONZE_TABLE,
-    comment="Bronze layer: Raw sensor data ingested from Kinesis",
-    table_properties={
-        "quality": "bronze",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.enableChangeDataFeed": "true",
-        "pipelines.reset.allowed": "false"
-    }
-)
-def raw_data():
-    """Bronze layer: Ingest raw Kinesis sensor data."""
-    
-    return (
-        spark.readStream
-        .format("kinesis")
-        .option("streamName", KINESIS_STREAM_NAME)
-        .option("initialPosition", "TRIM_HORIZON") 
-        .option("serviceCredential", SERVICE_CREDENTIAL_NAME)
-        .option("maxRecordsPerFetch", "10000")
-        .load()
-        .withColumn("ingestion_timestamp", F.current_timestamp())
-        # Capture Kinesis metadata
-        .withColumn("kinesis_sequence_number", F.col("sequenceNumber"))
-        .withColumn("kinesis_shard_id", F.col("shardId"))
-        .withColumn("kinesis_arrival_time", F.col("approximateArrivalTimestamp"))
-        .withColumn("partitionKey", F.col("partitionKey"))
-        # Parse data for basic extraction
-        .withColumn("parsed_data", F.from_json(F.col("data").cast("string"), sensor_schema))
-        .withColumn("ingest_date", F.to_date(F.col("ingestion_timestamp")))
-        # Basic experiment_id extraction for partitioning
-        .withColumn("experiment_id", F.coalesce(
-            F.regexp_extract(F.col("parsed_data.topic"), r"experiment/data_ingest/v1/([^/]+)/", 1),
-            F.regexp_extract(F.col("partitionKey"), r"/experiment/([^/]+)/", 1),
-            F.lit(None).cast(StringType())
-        ))
-        # clientid() from the IoT rule is the broker-authenticated Thing name (X.509 devices);
-        # extracted top-level to avoid evolving the non-resettable bronze parsed_data struct.
-        .withColumn("client_id", F.get_json_object(F.col("data").cast("string"), "$.client_id"))
-        .select(
-            "experiment_id",
-            "client_id",
-            "parsed_data",
-            "ingestion_timestamp",
-            "ingest_date",
-            "kinesis_sequence_number",
-            "kinesis_shard_id",
-            "kinesis_arrival_time",
-            "partitionKey"
-        )
-    )
-
-# COMMAND ----------
-
-# DBTITLE 1,Silver Layer - Clean Data
-@dlt.table(
-    name=SILVER_TABLE,
-    comment="Silver layer: Cleaned and standardized sensor data",
-    table_properties={
-        "quality": "silver",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.enableChangeDataFeed": "true"
-    }
-)
-@dlt.expect_or_drop("valid_timestamp", "timestamp IS NOT NULL")
-@dlt.expect("valid_device_id", "device_id IS NOT NULL")
-def clean_data():
-    """Silver layer: Clean and standardize sensor data, including imported data."""
-    bronze_df = dlt.read_stream(BRONZE_TABLE)
-    
-    df = (
-        bronze_df
-        .withColumn("device_id", F.col("parsed_data.device_id"))
-        .withColumn("device_name", F.col("parsed_data.device_name"))
-        .withColumn("device_version", F.col("parsed_data.device_version"))
-        .withColumn("device_battery", F.col("parsed_data.device_battery"))
-        .withColumn("device_firmware", F.col("parsed_data.device_firmware"))
-        .withColumn(
-            "sample",
-            decompress_sample(
-                F.col("parsed_data.sample"),
-                F.col("parsed_data._sample_encoding")
-            )
-        )
-        .withColumn("output", F.col("parsed_data.output"))
-        .withColumn("user_id", F.col("parsed_data.user_id"))
-        # Which protocol produced the row; topic = experiment/data_ingest/v1/<experiment>/<protocol>/...
-        .withColumn(
-            "protocol_id",
-            F.expr(r"nullif(regexp_extract(parsed_data.topic, 'experiment/data_ingest/v1/[^/]+/([^/]+)', 1), '')")
-        )
-        # Null on single-device uploads; rows of one multi-device workbook
-        # run share it and can be joined back on it.
-        .withColumn("workbook_run_id", F.col("parsed_data.workbook_run_id"))
-        # NOTE: timestamp === normalized UTC timestamp. timezone is the IANA name (e.g. "Europe/Amsterdam").
-        # Together they are the source of truth; all local-time representations are derived from these two.
-        .withColumn("timezone", F.col("parsed_data.timezone"))
-        .withColumn("timestamp", F.col("parsed_data.timestamp"))
-        .withColumn("processed_timestamp", F.current_timestamp())
-        .withColumn("date", F.to_date("timestamp"))
-        .withColumn("hour", F.hour("timestamp"))
-        .withColumn(
-            "ingest_latency_ms", 
-            F.unix_timestamp("ingestion_timestamp") - F.unix_timestamp("timestamp")
-        )
-    )
-
-    def _remap_macro_id(id_col):
-        """Remap legacy macro identifiers to actual UUIDs via chained CASE expression."""
-        result = id_col
-        for legacy_id, actual_id in LEGACY_MACRO_ID_MAP.items():
-            result = F.when(id_col == F.lit(legacy_id), F.lit(actual_id)).otherwise(result)
-        return result
-
-    df = df.withColumn(
-        "macros",
-        F.when(
-            F.col("parsed_data.macros").isNotNull(),
-            F.col("parsed_data.macros")
-        ).otherwise(
-            F.when(
-                F.col("sample").isNotNull(),
-                F.expr("""
-                    flatten(
-                        transform(
-                            from_json(sample, 'array<string>'),
-                            x -> transform(
-                                from_json(get_json_object(x, '$.macros'), 'array<string>'),
-                                m -> named_struct(
-                                    'id', m,
-                                    'name', m, 
-                                    'filename', m
-                                )
-                            )
-                        )
-                    )
-                """)
-            ).otherwise(F.array())
-        )
-    )
-
-    # Apply legacy macro ID remapping
-    df = df.withColumn(
-        "macros",
-        F.transform(
-            F.col("macros"),
-            lambda m: F.struct(
-                _remap_macro_id(m["id"]).alias("id"),
-                m["name"].alias("name"),
-                m["filename"].alias("filename")
-            )
-        )
-    )
-
-    df = df.withColumn(
-        "questions",
-        F.col("parsed_data.questions")
-    )
-    
-    df = df.withColumn(
-        "annotations",
-        F.coalesce(F.col("parsed_data.annotations"), F.array())
-    )
-
-    df = df.withColumn(
-        "id",
-        F.abs(
-            F.hash(
-                F.col("experiment_id"),
-                F.col("device_id"),
-                F.col("timestamp"),
-                F.col("sample"),
-                F.col("ingestion_timestamp"),
-                F.col("kinesis_sequence_number")
-            )
-        )
-    )
-
-    # Populate missing annotation IDs and rowIds
-    # If annotations come from payload without IDs, generate them here
-    df = df.withColumn(
-        "annotations",
-        F.expr("""
-            transform(annotations, a -> struct(
-                coalesce(a.id, uuid()) as id,
-                coalesce(a.rowId, cast(id as string)) as rowId,
-                a.type as type,
-                a.content as content,
-                a.createdBy as createdBy,
-                a.createdByName as createdByName,
-                coalesce(a.createdAt, current_timestamp()) as createdAt,
-                coalesce(a.updatedAt, current_timestamp()) as updatedAt
-            ))
-        """)
-    )
-
-    df = df.withColumn("skip_macro_processing", F.lit(None).cast("boolean"))
-    
-    # Select final columns for bronze-sourced silver data
-    bronze_clean = df.select(
-        "id",
-        "device_id",
-        "client_id",
-        "device_name",
-        "device_version",
-        "device_battery",
-        "device_firmware",
-        "sample",
-        "output",
-        "macros",
-        "questions",
-        "annotations",
-        "user_id",
-        "timezone",
-        "experiment_id",
-        "protocol_id",
-        "workbook_run_id",
-        "timestamp",
-        "date",
-        "hour",
-        "ingest_latency_ms",
-        "processed_timestamp",
-        "skip_macro_processing"
-    )
-
-    # Read imported data and align to the same schema
-    imported_df = dlt.read_stream(RAW_IMPORTED_DATA_TABLE)
-
-    imported_clean = (
-        imported_df
-        .withColumn("id", F.col("id").cast("long"))
-        .withColumn("processed_timestamp", F.current_timestamp())
-        .withColumn("date", F.to_date("timestamp"))
-        .withColumn("hour", F.hour("timestamp"))
-        .withColumn("ingest_latency_ms", F.lit(None).cast("long"))
-        .withColumn("timezone", F.lit(None).cast("string"))
-        # Imported/transfer rows have no MQTT client identity.
-        .withColumn("client_id", F.lit(None).cast("string"))
-        # Build macros array from macro columns (empty when no macro)
-        # Apply legacy macro ID remapping inline
-        .withColumn(
-            "macros",
-            F.when(
-                F.col("macro_id").isNotNull(),
-                F.array(
-                    F.struct(
-                        _remap_macro_id(F.col("macro_id")).alias("id"),
-                        F.coalesce(F.col("macro_name"), F.col("macro_filename"), F.col("macro_id")).alias("name"),
-                        F.coalesce(F.col("macro_filename"), F.col("macro_id")).alias("filename")
-                    )
-                )
-            ).otherwise(F.array())
-        )
-        # Parse questions from JSON string if present, otherwise empty array
-        .withColumn(
-            "questions",
-            F.when(
-                F.col("questions").isNotNull(),
-                F.from_json(F.col("questions"), ArrayType(question_schema))
-            ).otherwise(F.array())
-        )
-        .withColumn("annotations", F.array())
-        # Imported/legacy data has no topic and no workbook-run concept.
-        .withColumn("protocol_id", F.lit(None).cast("string"))
-        .withColumn("workbook_run_id", F.lit(None).cast("string"))
-        # Mark imported data to skip macro processing
-        .withColumn("skip_macro_processing", F.lit(True))
-        .select(
-            "id",
-            "device_id",
-            "client_id",
-            "device_name",
-            "device_version",
-            "device_battery",
-            "device_firmware",
-            "sample",
-            "output",
-            "macros",
-            "questions",
-            "annotations",
-            "user_id",
-            "timezone",
-            "experiment_id",
-            "protocol_id",
-            "workbook_run_id",
-            "timestamp",
-            "date",
-            "hour",
-            "ingest_latency_ms",
-            "processed_timestamp",
-            "skip_macro_processing"
-        )
-    )
-
-    # Union bronze-sourced data with imported data
-    return bronze_clean.unionByName(imported_clean)
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Experiment Status
-@dlt.table(
-    name=EXPERIMENT_STATUS_TABLE,
-    comment="Gold layer: Materialized view tracking experiment freshness status",
-    table_properties={
-        "quality": "gold",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true"
-    }
-)
-def experiment_status():
-    """Track experiment data freshness."""
-    FRESHNESS_THRESHOLD_MINUTES = 60
-    
-    # Read from silver table
-    silver_df = dlt.read(SILVER_TABLE)
-    
-    # Get the current timestamp for comparison
-    current_timestamp = F.current_timestamp()
-    
-    # Calculate the latest timestamp for each experiment_id
-    experiment_status_df = (
-        silver_df
-        .groupBy("experiment_id")
-        .agg(
-            F.max("timestamp").alias("latest_timestamp"),
-            F.max("processed_timestamp").alias("latest_processed_timestamp")
-        )
-        .filter("experiment_id IS NOT NULL")
-    )
-    
-    freshness_threshold_seconds = FRESHNESS_THRESHOLD_MINUTES * 60
-    status_df = (
-        experiment_status_df
-        .withColumn(
-            "status",
-            F.when(
-                (current_timestamp.cast("long") - F.col("latest_processed_timestamp").cast("long")) <= freshness_threshold_seconds,
-                F.lit("fresh")
-            ).otherwise(F.lit("stale"))
-        )
-        .withColumn("status_updated_at", current_timestamp)
-    )
-    
-    return status_df.select(
-        "experiment_id",
-        "latest_timestamp",
-        "latest_processed_timestamp",
-        "status",
-        "status_updated_at"
-    )
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Experiment Raw Data
-@dlt.table(
-    name=EXPERIMENT_RAW_DATA_TABLE,
-    comment="Gold layer: Per-experiment raw sample data partitioned by experiment_id with VARIANT sample",
-    table_properties={
-        "quality": "gold",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.enableRowTracking": "true",
-        "delta.enableChangeDataFeed": "true",
-        "delta.feature.variantType-preview": "supported"
-    }
-)
-def experiment_raw_data():
-    """Per-experiment raw sample data with VARIANT support."""
-    
-    # Define UDF for sanitizing question labels
-    @F.pandas_udf(ArrayType(StructType([
-        StructField("question_label", StringType(), True),
-        StructField("question_answer", StringType(), True)
-    ])))
-    def sanitize_questions_udf(questions: pd.Series) -> pd.Series:
-        
-        def sanitize_label(label):
-            if not label:
-                return "question_empty"
-
-            # Allowlist: lowercase ASCII letters, digits, underscore.
-            # Anything else collapses to a single underscore so the result is
-            # safe as a Spark identifier and as a JSON/VARIANT key, regardless
-            # of where it gets rendered downstream (SQL, CSV, Excel, UI).
-            sanitized = ''.join(
-                c if (c == '_' or (c.isascii() and c.isalnum())) else '_'
-                for c in label.lower()
-            )
-            while '__' in sanitized:
-                sanitized = sanitized.replace('__', '_')
-            sanitized = sanitized.strip('_')
-
-            if not sanitized or sanitized[0].isdigit():
-                sanitized = f"question_{sanitized}"
-
-            return sanitized
-        
-        def sanitize_questions_array(questions_array):
-            if questions_array is None or len(questions_array) == 0:
-                return []
-
-            # Disambiguate duplicate sanitized labels so downstream
-            # map_from_arrays doesn't collapse them and lose answers.
-            # Strategy:
-            #   - unique label in row -> bare label
-            #   - label collides, sanitized text differs from sanitized label
-            #       -> "<label>__<text>"
-            #   - label collides but text adds nothing (empty, or same as label
-            #       after sanitization, e.g. uncustomized default)
-            #       -> positional fallback "_2", "_3" on bare label
-            #   - label and text both collide
-            #       -> positional fallback on the combined "<label>__<text>" key
-            base_labels = [
-                sanitize_label(q.get('question_label')) if q else None
-                for q in questions_array
-            ]
-            label_total: dict[str, int] = {}
-            for bl in base_labels:
-                if bl is not None:
-                    label_total[bl] = label_total.get(bl, 0) + 1
-
-            key_counts: dict[str, int] = {}
-            result = []
-            for i, q in enumerate(questions_array):
-                if not q:
-                    continue
-                base = base_labels[i]
-                if label_total.get(base, 0) <= 1:
-                    key = base
-                else:
-                    text = q.get('question_text')
-                    text_part = sanitize_label(text) if text else None
-                    if not text_part or text_part == base:
-                        candidate = base
-                    else:
-                        candidate = f"{base}__{text_part}"
-                    count = key_counts.get(candidate, 0) + 1
-                    key_counts[candidate] = count
-                    key = candidate if count == 1 else f"{candidate}_{count}"
-                result.append({
-                    'question_label': key,
-                    'question_answer': q.get('question_answer')
-                })
-            return result
-
-        return questions.apply(sanitize_questions_array)
-    
-    return (
-        dlt.read_stream(SILVER_TABLE)
-        .filter("experiment_id IS NOT NULL")
-        .withColumn("data", F.expr("parse_json(sample)"))
-        .withColumn("output_data", F.expr("try_parse_json(output)"))
-        .withColumn(
-            "questions_sanitized",
-            F.when(
-                F.col("questions").isNotNull() & (F.size("questions") > 0),
-                sanitize_questions_udf(F.col("questions"))
-            )
-        )
-        .withColumn(
-            "questions_data",
-            F.when(
-                F.col("questions_sanitized").isNotNull() & (F.size("questions_sanitized") > 0),
-                F.expr("""
-                    parse_json(
-                        to_json(
-                            map_from_arrays(
-                                transform(questions_sanitized, q -> q.question_label),
-                                transform(questions_sanitized, q -> q.question_answer)
-                            )
-                        )
-                    )
-                """)
-            )
-        )
-        .select(
-            "id",
-            "experiment_id",
-            "device_id",
-            "client_id",
-            "device_name",
-            "timestamp",
-            "timezone",
-            "macros",
-            "questions_data",
-            "annotations",
-            "user_id",
-            "protocol_id",
-            "workbook_run_id",
-            "data",
-            "output_data",
-            "date",
-            "processed_timestamp",
-            "skip_macro_processing"
-        )
-        .transform(lambda df: apply_inline_repairs(df, EXPERIMENT_RAW_DATA_TABLE))
-    )
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Experiment Device Data
-@dlt.table(
-    name=EXPERIMENT_DEVICE_DATA_TABLE,
-    comment="Gold layer: Device metadata aggregated per experiment",
-    table_properties={
-        "quality": "gold",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-    }
-)
-def experiment_device_data():
-    """
-    Aggregate device stats per experiment from clean_data.
-    """
-    silver_df = dlt.read(SILVER_TABLE)
-    devices = dlt.read(EXPERIMENT_DEVICES_TABLE)
-
-    aggregated = (
-        silver_df
-        .filter("experiment_id IS NOT NULL")
-        .groupBy("experiment_id", "device_id", "device_firmware")
-        .agg(
-            F.max("device_name").alias("device_name"),
-            F.max("device_version").alias("device_version"),
-            F.max("device_battery").alias("device_battery"),
-            F.max("client_id").alias("client_id"),
-            F.count("*").alias("total_measurements"),
-            F.max("processed_timestamp").alias("processed_timestamp")
-        )
-        .withColumn(
-            "id",
-            F.abs(
-                F.hash(
-                    F.col("experiment_id"),
-                    F.col("device_id"),
-                    F.col("device_firmware")
-                )
-            )
-        )
-    )
-
-    # Attach the registry-resolved device struct via the trusted client_id
-    # (NULL for Cognito/unregistered rows — left join keeps every device row).
-    return (
-        aggregated
-        .join(
-            devices,
-            (aggregated.experiment_id == devices.experiment_id)
-            & (aggregated.client_id == devices.client_id),
-            "left"
-        )
-        .select(
-            aggregated.id,
-            aggregated.experiment_id,
-            aggregated.device_id,
-            aggregated.client_id,
-            aggregated.device_firmware,
-            aggregated.device_name,
-            aggregated.device_version,
-            aggregated.device_battery,
-            aggregated.total_measurements,
-            aggregated.processed_timestamp,
-            devices.device,
-        )
-    )
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Experiment Macro Data
-@dlt.table(
-    name=EXPERIMENT_MACRO_DATA_TABLE,
-    comment="Gold layer: Unified macro processing with VARIANT column for flexible schema",
-    table_properties={
-        "quality": "gold",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.enableRowTracking": "true",
-        "delta.enableChangeDataFeed": "true",
-        "delta.feature.variantType-preview": "supported"
-    }
-)
-def experiment_macro_data():
-    """Process macros with VARIANT output column."""
-
-    sandbox_macro_udf = make_execute_macro_udf(ENVIRONMENT, dbutils)
-
-    base_df = (
-        dlt.read_stream(EXPERIMENT_RAW_DATA_TABLE)
-        .filter("macros IS NOT NULL")
-        .filter("size(macros) > 0")
-        .select(
-            "id",
-            "experiment_id",
-            "device_id",
-            "client_id",
-            "device_name",
-            "timestamp",
-            "timezone",
-            "user_id",
-            "data",
-            "output_data",
-            "date",
-            "processed_timestamp",
-            "questions_data",
-            "annotations",
-            "skip_macro_processing",
-            F.explode("macros").alias("macro")
-        )
-        .select(
-            "id",
-            "experiment_id",
-            "device_id",
-            "client_id",
-            "device_name",
-            "timestamp",
-            "timezone",
-            "user_id",
-            "data",
-            "output_data",
-            "date",
-            "processed_timestamp",
-            "questions_data",
-            "annotations",
-            "skip_macro_processing",
-            F.col("macro.id").alias("macro_id"),
-            F.col("macro.name").alias("macro_name"),
-            F.col("macro.filename").alias("macro_filename")
-        )
-    )
-
-    return (
-        base_df
-        .transform(lambda df: apply_inline_repairs(df, EXPERIMENT_MACRO_DATA_TABLE))
-        # NULL.rlike(...) returns NULL (treated as false in F.when), so the
-        # explicit isNotNull() guard is required, otherwise null macro_ids
-        # would silently land with no output and no error.
-        .withColumn(
-            "sandbox_result",
-            F.when(
-                F.col("macro_id").isNotNull()
-                & F.col("macro_id").rlike(MACRO_ID_UUID_PATTERN)
-                & ~F.coalesce(F.col("skip_macro_processing"), F.lit(False)),
-                sandbox_macro_udf(
-                    F.struct("id", "macro_id", F.col("data"))
-                ),
-            )
-        )
-        # For imported rows, use pre-computed output_data; otherwise use UDF result
-        .withColumn(
-            "macro_output",
-            F.when(
-                F.col("skip_macro_processing") == True,
-                F.col("output_data")
-            ).otherwise(
-                F.when(
-                    F.col("sandbox_result.result").isNotNull(),
-                    F.expr("parse_json(sandbox_result.result)"),
-                )
-            ),
-        )
-        .withColumn(
-            "macro_error",
-            F.when(
-                F.col("skip_macro_processing") == True,
-                F.lit(None).cast("string")
-            ).when(
-                F.col("macro_id").isNull(),
-                F.lit("Invalid macro_id (null)")
-            ).when(
-                ~F.col("macro_id").rlike(MACRO_ID_UUID_PATTERN),
-                F.concat(F.lit("Invalid macro_id (not UUID): "), F.col("macro_id"))
-            ).otherwise(F.col("sandbox_result.error"))
-        )
-        .withColumn(
-            "macro_row_id",
-            F.abs(
-                F.hash(
-                    F.col("id"),
-                    F.col("macro_filename"),
-                    F.col("processed_timestamp"),
-                )
-            ),
-        )
-        .select(
-            "experiment_id",
-            F.col("macro_row_id").alias("id"),
-            F.col("id").alias("raw_id"),
-            "device_id",
-            "client_id",
-            "device_name",
-            "timestamp",
-            "timezone",
-            "user_id",
-            "macro_id",
-            "macro_name",
-            "macro_filename",
-            "macro_output",
-            "macro_error",
-            "processed_timestamp",
-            "date",
-            "questions_data",
-            "annotations",
-        )
-    )
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Experiment Table Metadata
-@dlt.table(
-    name=EXPERIMENT_TABLE_METADATA,
-    comment="Gold layer: Consolidated metadata cache for all experiment tables (row counts, schemas). Single query optimization. Replaces EXPERIMENT_MACROS_TABLE and EXPERIMENT_QUESTIONS_TABLE.",
-    table_properties={
-        "quality": "gold",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.feature.variantType-preview": "supported",
-    }
-)
-def experiment_table_metadata():
-    """Metadata for all experiment tables."""
-
-    # Pre-compute custom_metadata_schema per experiment from the metadata source.
-    # Each experiment's metadata blob has $.rows, an array of VARIANT objects
-    # whose keys are already human-readable (remapped at upload time).
-    # We drop internal keys (_id, identifier column) from each row before schema
-    # inference so the backend doesn't expand them into duplicate query columns.
-    metadata_source = dlt.read(METADATA_SOURCE_TABLE)
-    custom_metadata_schemas = (
-        metadata_source
-        .select(
-            F.col("experiment_id"),
-            F.expr("variant_get(metadata, '$.identifierColumnId', 'STRING')").alias("_id_col"),
-            F.expr("explode(variant_get(metadata, '$.rows', 'ARRAY<VARIANT>'))").alias("_row"),
-        )
-        .withColumn("_row", F.expr("""
-            parse_json(to_json(map_filter(
-                cast(_row AS MAP<STRING, VARIANT>),
-                (k, v) -> k != '_id' AND k != _id_col
-            )))
-        """))
-        .groupBy("experiment_id")
-        .agg(
-            F.expr("nullif(schema_of_variant_agg(_row), 'VOID')").alias("custom_metadata_schema")
-        )
-    )
-
-    macro_metadata = (
-        dlt.read(EXPERIMENT_MACRO_DATA_TABLE)
-        .filter("macro_output IS NOT NULL")
-        .groupBy("experiment_id", "macro_id")
-        .agg(
-            F.count("*").alias("row_count"),
-            F.expr("nullif(schema_of_variant_agg(macro_output), 'VOID')").alias("macro_schema"),
-            F.expr("nullif(schema_of_variant_agg(questions_data), 'VOID')").alias("questions_schema")
-        )
-        .join(custom_metadata_schemas, "experiment_id", "left")
-        .select(
-            F.col("experiment_id"),
-            F.col("macro_id").alias("identifier"),
-            F.lit("macro").alias("table_type"),
-            F.lit(None).cast("string").alias("display_name"),
-            F.col("row_count"),
-            F.col("macro_schema"),
-            F.col("questions_schema"),
-            F.col("custom_metadata_schema"),
-            F.lit(None).cast("string").alias("upload_schema"),
-        )
-    )
-
-    raw_data_metadata = (
-        dlt.read(EXPERIMENT_RAW_DATA_TABLE)
-        .groupBy("experiment_id")
-        .agg(
-            F.count("*").alias("row_count"),
-            F.expr("nullif(schema_of_variant_agg(questions_data), 'VOID')").alias("questions_schema")
-        )
-        .join(custom_metadata_schemas, "experiment_id", "left")
-        .select(
-            F.col("experiment_id"),
-            F.lit("raw_data").alias("identifier"),
-            F.lit("static").alias("table_type"),
-            F.lit(None).cast("string").alias("display_name"),
-            F.col("row_count"),
-            F.lit(None).cast("string").alias("macro_schema"),
-            F.col("questions_schema"),
-            F.col("custom_metadata_schema"),
-            F.lit(None).cast("string").alias("upload_schema"),
-        )
-    )
-
-    device_metadata = (
-        dlt.read(EXPERIMENT_DEVICE_DATA_TABLE)
-        .groupBy("experiment_id")
-        .agg(F.count("*").alias("row_count"))
-        .select(
-            F.col("experiment_id"),
-            F.lit("device").alias("identifier"),
-            F.lit("static").alias("table_type"),
-            F.lit(None).cast("string").alias("display_name"),
-            F.col("row_count"),
-            F.lit(None).cast("string").alias("macro_schema"),
-            F.lit(None).cast("string").alias("questions_schema"),
-            F.lit(None).cast("string").alias("custom_metadata_schema"),
-            F.lit(None).cast("string").alias("upload_schema"),
-        )
-    )
-
-    upload_metadata = (
-        dlt.read(EXPERIMENT_UPLOADED_DATA_TABLE)
-        .filter("uploaded_data IS NOT NULL")
-        .groupBy("experiment_id", "upload_table_id")
-        .agg(
-            F.count("*").alias("row_count"),
-            # Latest user-chosen label for this logical upload table, picked by
-            # the most recent uploaded_at so renames take effect deterministically.
-            F.max(F.struct(F.col("uploaded_at"), F.col("upload_table_name"))).getField("upload_table_name").alias("display_name"),
-            F.expr("nullif(schema_of_variant_agg(uploaded_data), 'VOID')").alias("upload_schema")
-        )
-        .select(
-            F.col("experiment_id"),
-            F.col("upload_table_id").alias("identifier"),
-            F.lit("upload").alias("table_type"),
-            F.col("display_name"),
-            F.col("row_count"),
-            F.lit(None).cast("string").alias("macro_schema"),
-            F.lit(None).cast("string").alias("questions_schema"),
-            F.lit(None).cast("string").alias("custom_metadata_schema"),
-            F.col("upload_schema"),
-        )
-    )
-
-    return (
-        macro_metadata
-        .unionByName(raw_data_metadata)
-        .unionByName(device_metadata)
-        .unionByName(upload_metadata)
-    )
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Experiment Contributors
-@dlt.table(
-    name=EXPERIMENT_CONTRIBUTORS_TABLE,
-    comment="Gold layer: Cached user profiles for enrichment (full refresh on each pipeline run)",
-    table_properties={
-        "quality": "gold",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.enableRowTracking": "true",
-        "delta.enableChangeDataFeed": "true",
-    }
-)
-def experiment_contributors():
-    """Cached user profiles per experiment, keyed by (experiment_id, user_id).
-
-    Sourced from sensor measurements plus data uploaders: an uploader may never
-    have submitted a measurement, so include their created_by here too, otherwise
-    the enriched_experiment_uploaded_data contributor join can't resolve them.
-    """
-    sensor_users = (
-        dlt.read(SILVER_TABLE)
-        .filter("experiment_id IS NOT NULL")
-        .filter("user_id IS NOT NULL")
-        .select("experiment_id", "user_id")
-    )
-
-    upload_users = (
-        dlt.read(EXPERIMENT_UPLOADED_DATA_TABLE)
-        .filter("experiment_id IS NOT NULL")
-        .filter("created_by IS NOT NULL")
-        .select("experiment_id", F.col("created_by").alias("user_id"))
-    )
-
-    unique_users = sensor_users.unionByName(upload_users).distinct()
-
-    return add_user_column(unique_users, ENVIRONMENT, dbutils)
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Experiment Devices
-@dlt.table(
-    name=EXPERIMENT_DEVICES_TABLE,
-    comment="Gold layer: Registry-resolved devices per experiment (client_id -> device), full refresh each run.",
-    table_properties={
-        "quality": "gold",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.enableRowTracking": "true",
-        "delta.enableChangeDataFeed": "true",
-    }
-)
-def experiment_devices():
-    """Devices observed per experiment, resolved against the registry via the
-    broker-authenticated client_id (== Thing name for X.509 devices). Cognito
-    rows have a non-Thing client_id and resolve to a NULL device struct.
-
-    Mirrors experiment_contributors: distinct keys from the data, enriched once
-    per run from the backend, then joined by the gold device dimension.
-    """
-    unique_devices = (
-        dlt.read(SILVER_TABLE)
-        .filter("experiment_id IS NOT NULL")
-        .filter("client_id IS NOT NULL")
-        .select("experiment_id", "client_id")
-        .distinct()
-    )
-
-    return add_device_registry(unique_devices, ENVIRONMENT, dbutils)
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Annotations Source (DLT mirror)
-@dlt.table(
-    name=ANNOTATIONS_SOURCE_TABLE,
-    comment="Gold layer: DLT-tracked mirror of the experiment_annotations table for incremental refresh support.",
-    table_properties={
-        "quality": "gold",
-        "delta.enableRowTracking": "true",
-        "delta.enableChangeDataFeed": "true",
-    }
-)
-def experiment_annotations_source():
-    """DLT mirror of the backend-managed experiment_annotations table."""
-    return spark.read.table(f"{CATALOG_NAME}.centrum.experiment_annotations")
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Metadata Source (DLT mirror)
-@dlt.table(
-    name=METADATA_SOURCE_TABLE,
-    comment="Gold layer: DLT-tracked mirror of the experiment_metadata table for incremental refresh support.",
-    table_properties={
-        "quality": "gold",
-        "delta.enableRowTracking": "true",
-        "delta.enableChangeDataFeed": "true",
-        "delta.feature.variantType-preview": "supported",
-    }
-)
-def experiment_metadata_source():
-    """DLT mirror of the backend-managed experiment_metadata table."""
-    return spark.read.table(f"{CATALOG_NAME}.centrum.experiment_custom_metadata")
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Enriched Experiment Raw Data
-@dlt.table(
-    name=ENRICHED_RAW_DATA_VIEW,
-    comment="Enriched materialized view: Raw data with questions, user struct, and annotations. Qualified for incremental refresh.",
-    table_properties={
-        "quality": "gold",
-        "delta.enableRowTracking": "true",
-        "delta.enableChangeDataFeed": "true",
-        "delta.enableDeletionVectors": "true",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.feature.variantType-preview": "supported",
-    }
-)
-def enriched_experiment_raw_data():
-    """Enriched raw data with user profiles, annotations, and user metadata."""
-    raw_data = dlt.read(EXPERIMENT_RAW_DATA_TABLE)
-    contributors = dlt.read(EXPERIMENT_CONTRIBUTORS_TABLE)
-    devices = dlt.read(EXPERIMENT_DEVICES_TABLE)
-    annotations_source = dlt.read(ANNOTATIONS_SOURCE_TABLE)
-    metadata_source = dlt.read(METADATA_SOURCE_TABLE)
-
-    enriched = (
-        raw_data
-        .join(
-            contributors,
-            (raw_data.experiment_id == contributors.experiment_id) &
-            (raw_data.user_id == contributors.user_id),
-            "left"
-        )
-        .join(
-            devices,
-            (raw_data.experiment_id == devices.experiment_id) &
-            (raw_data.client_id == devices.client_id),
-            "left"
-        )
-        .select(
-            raw_data.experiment_id,
-            raw_data.id,
-            raw_data.device_id,
-            raw_data.device_name,
-            raw_data.timestamp,  # kept for backwards compatibility (downstream consumers order by timestamp)
-            raw_data.timestamp.alias("measurement_time_utc"),
-            raw_data.timezone,
-            raw_data.date,
-            raw_data.macros,
-            raw_data.questions_data,
-            raw_data.annotations,
-            contributors.user.alias("contributor"),
-            devices.device.alias("device"),
-            raw_data.protocol_id,
-            raw_data.workbook_run_id,
-            raw_data.data,
-            raw_data.processed_timestamp
-        )
-        .withColumn(
-            "measurement_time_local",
-            F.when(
-                F.col("timezone").isNotNull(),
-                F.date_format(F.from_utc_timestamp(F.col("measurement_time_utc"), F.col("timezone")), "yyyy-MM-dd HH:mm:ss")
-            )
-        )
-        .withColumn(
-            "local_time",
-            F.when(
-                F.col("timezone").isNotNull(),
-                F.date_format(F.from_utc_timestamp(F.col("measurement_time_utc"), F.col("timezone")), "HH:mm")
-            )
-        )
-    )
-    
-    enriched = add_annotation_column(enriched, annotations_source)
-    
-    return add_custom_metadata_column(enriched, metadata_source)
-
-
-# COMMAND ----------
-
-# DBTITLE 1,Gold Layer - Enriched Experiment Macro Data
-@dlt.table(
-    name=ENRICHED_MACRO_DATA_VIEW,
-    comment="Enriched materialized view: Macro data with expanded VARIANT, questions, user struct, and annotations. Qualified for incremental refresh.",
-    table_properties={
-        "quality": "gold",
-        "delta.enableRowTracking": "true",
-        "delta.enableChangeDataFeed": "true",
-        "delta.enableDeletionVectors": "true",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.feature.variantType-preview": "supported",
-    }
-)
-def enriched_experiment_macro_data():
-    """Enriched macro data with user profiles, annotations, and user metadata."""
-    macro_data = dlt.read(EXPERIMENT_MACRO_DATA_TABLE)
-    contributors = dlt.read(EXPERIMENT_CONTRIBUTORS_TABLE)
-    devices = dlt.read(EXPERIMENT_DEVICES_TABLE)
-    annotations_source = dlt.read(ANNOTATIONS_SOURCE_TABLE)
-    metadata_source = dlt.read(METADATA_SOURCE_TABLE)
-
-    enriched = (
-        macro_data
-        .join(
-            contributors,
-            (macro_data.experiment_id == contributors.experiment_id) &
-            (macro_data.user_id == contributors.user_id),
-            "left"
-        )
-        .join(
-            devices,
-            (macro_data.experiment_id == devices.experiment_id) &
-            (macro_data.client_id == devices.client_id),
-            "left"
-        )
-        .select(
-            macro_data.experiment_id,
-            macro_data.id,
-            macro_data.raw_id,
-            macro_data.device_id,
-            macro_data.device_name,
-            macro_data.timestamp,  # kept for backwards compatibility (downstream consumers order by timestamp)
-            macro_data.timestamp.alias("measurement_time_utc"),
-            macro_data.timezone,
-            macro_data.date,
-            contributors.user.alias("contributor"),
-            devices.device.alias("device"),
-            macro_data.macro_id,
-            macro_data.macro_name,
-            macro_data.macro_filename,
-            macro_data.macro_output,
-            macro_data.macro_error,
-            macro_data.processed_timestamp,
-            macro_data.questions_data,
-            macro_data.annotations
-        )
-        .withColumn(
-            "measurement_time_local",
-            F.when(
-                F.col("timezone").isNotNull(),
-                F.date_format(F.from_utc_timestamp(F.col("measurement_time_utc"), F.col("timezone")), "yyyy-MM-dd HH:mm:ss")
-            )
-        )
-        .withColumn(
-            "local_time",
-            F.when(
-                F.col("timezone").isNotNull(),
-                F.date_format(F.from_utc_timestamp(F.col("measurement_time_utc"), F.col("timezone")), "HH:mm")
-            )
-        )
-    )
-    
-    enriched = add_annotation_column(enriched, annotations_source)
-    
-    return add_custom_metadata_column(enriched, metadata_source)
-
-# COMMAND ----------
-
-# DBTITLE 1,Raw Imported Data - Streaming Table
-@dlt.table(
-    name=RAW_IMPORTED_DATA_TABLE,
-    comment="Streaming table: Imported measurement data from external platforms (e.g., PhotosynQ transfers), partitioned by experiment_id",
-    table_properties={
-        "quality": "bronze",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true"
-    },
-    partition_cols=["experiment_id"]
-)
-def raw_imported_data():
-    """Streaming ingestion of imported measurement data from parquet files in data-imports volume."""
-    imported_path = f"/Volumes/{CATALOG_NAME}/centrum/data-imports/*/photosynq_transfer"
-    
-    imported_data_schema = StructType([
-        StructField("id", StringType(), True),
-        StructField("device_id", StringType(), True),
-        StructField("device_name", StringType(), True),
-        StructField("device_version", StringType(), True),
-        StructField("device_battery", DoubleType(), True),
-        StructField("device_firmware", StringType(), True),
-        StructField("sample", StringType(), True),
-        StructField("output", StringType(), True),
-        StructField("user_id", StringType(), True),
-        StructField("experiment_id", StringType(), True),
-        StructField("timestamp", TimestampType(), True),
-        StructField("macro_id", StringType(), True),
-        StructField("macro_filename", StringType(), True),
-        StructField("macro_name", StringType(), True),
-        StructField("questions", StringType(), True),
-    ])
-    
-    df = (
-        spark.readStream
-        .format("cloudFiles")
-        .option("cloudFiles.format", "parquet")
-        .option("recursiveFileLookup", "true")
-        .option("ignoreMissingFiles", "true")
-        .schema(imported_data_schema)
-        .load(imported_path)
-    )
-    
-    return (
-        df
-        .withColumn(
-            "id",
-            F.coalesce(
-                F.col("id"),
-                F.abs(F.hash(*[F.col(c) for c in df.columns])).cast("string")
-            )
-        )
-        .withColumn("ingestion_timestamp", F.current_timestamp())
-    )
-
-# COMMAND ----------
-
-# DBTITLE 1,Raw Uploaded Data - Streaming Table
-@dlt.table(
-    name=RAW_UPLOADED_DATA_TABLE,
-    comment="Streaming table: User-uploaded data with variant-encoded row payloads, partitioned by experiment_id",
-    table_properties={
-        "quality": "bronze",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.feature.variantType-preview": "supported",
-    },
-    partition_cols=["experiment_id"]
-)
-def raw_uploaded_data():
-    """Bronze: streaming ingestion of user-uploaded tabular data from parquet files in the data-imports volume.
-
-    The Python upload task writes uploaded_data as a JSON string in parquet, plus the stable upload_table_id
-    that anchors a logical upload table across batches. Here we parse uploaded_data into VARIANT so downstream
-    gold derivation and metadata aggregation can infer per-table schemas via schema_of_variant_agg.
-    Peer of raw_imported_data."""
-    uploaded_path = f"/Volumes/{CATALOG_NAME}/centrum/data-imports/*/processed-uploads"
-
-    uploaded_data_schema = StructType([
-        StructField("experiment_id", StringType(), True),
-        StructField("upload_table_id", StringType(), True),
-        StructField("upload_table_name", StringType(), True),
-        StructField("upload_id", StringType(), True),
-        StructField("created_by", StringType(), True),
-        StructField("uploaded_at", TimestampType(), True),
-        StructField("uploaded_data", StringType(), True),
-        StructField("row_index", LongType(), True),
-    ])
-
-    return (
-        spark.readStream
-        .format("cloudFiles")
-        .option("cloudFiles.format", "parquet")
-        .option("recursiveFileLookup", "true")
-        .option("ignoreMissingFiles", "true")
-        .schema(uploaded_data_schema)
-        .load(uploaded_path)
-        .withColumn("uploaded_data", F.expr("try_parse_json(uploaded_data)"))
-    )
-
-# COMMAND ----------
-
-# DBTITLE 1,Experiment Uploaded Data - Gold
-@dlt.table(
-    name=EXPERIMENT_UPLOADED_DATA_TABLE,
-    comment="Gold layer: User-uploaded tabular data with VARIANT row payloads. Peer of experiment_macro_data; anchored by stable upload_table_id.",
-    table_properties={
-        "quality": "gold",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.enableRowTracking": "true",
-        "delta.enableChangeDataFeed": "true",
-        "delta.feature.variantType-preview": "supported",
-    },
-)
-def experiment_uploaded_data():
-    """Gold: rows from raw_uploaded_data with a per-row id, ready for the enriched view.
-
-    id is an opaque hash of (upload_id, row_index), mirroring experiment_macro_data's
-    derived id; annotations join on it. xxhash64 rather than 32-bit hash() because a
-    single upload can exceed the ~77k rows where 32-bit collisions become likely.
-    """
-    return (
-        dlt.read_stream(RAW_UPLOADED_DATA_TABLE)
-        .withColumn(
-            "id",
-            F.abs(F.xxhash64(F.col("upload_id"), F.col("row_index"))),
-        )
-        .select(
-            F.col("id"),
-            F.col("experiment_id"),
-            F.col("upload_table_id"),
-            F.col("upload_table_name"),
-            F.col("upload_id"),
-            F.col("created_by"),
-            F.col("uploaded_at"),
-            F.col("uploaded_data"),
-        )
-    )
-
-# COMMAND ----------
-
-# DBTITLE 1,Enriched Experiment Uploaded Data
-@dlt.table(
-    name=ENRICHED_UPLOADED_DATA_VIEW,
-    comment="Enriched view: uploaded rows joined with contributors, annotations, and experiment metadata. Mirrors enriched_experiment_macro_data.",
-    table_properties={
-        "quality": "gold",
-        "delta.enableRowTracking": "true",
-        "delta.enableChangeDataFeed": "true",
-        "delta.enableDeletionVectors": "true",
-        "pipelines.autoOptimize.managed": "true",
-        "delta.autoOptimize.optimizeWrite": "true",
-        "delta.autoOptimize.autoCompact": "true",
-        "delta.feature.variantType-preview": "supported",
-    },
-)
-def enriched_experiment_uploaded_data():
-    uploaded = dlt.read(EXPERIMENT_UPLOADED_DATA_TABLE)
-    contributors = dlt.read(EXPERIMENT_CONTRIBUTORS_TABLE)
-    annotations_source = dlt.read(ANNOTATIONS_SOURCE_TABLE)
-    metadata_source = dlt.read(METADATA_SOURCE_TABLE)
-
-    enriched = (
-        uploaded
-        .join(
-            contributors,
-            (uploaded.experiment_id == contributors.experiment_id)
-            & (uploaded.created_by == contributors.user_id),
-            "left",
-        )
-        .select(
-            uploaded.id,
-            uploaded.experiment_id,
-            uploaded.upload_table_id,
-            uploaded.upload_table_name,
-            uploaded.upload_id,
-            uploaded.uploaded_at,
-            uploaded.uploaded_data,
-            contributors.user.alias("contributor"),
-        )
-    )
-
-    enriched = add_annotation_column(enriched, annotations_source)
-
-    return add_custom_metadata_column(enriched, metadata_source)
-
-# COMMAND ----------
-
-# DBTITLE 1,Event Hook - Slack Notifications
-@dlt.on_event_hook(max_allowable_consecutive_failures=3)
-def send_slack_notifications(event):
-
-    # Get the webhook URL from the secret scope
-    SLACK_WEBHOOK_URL = dbutils.secrets.get(scope=f"event-hooks-{ENVIRONMENT}", key="slack-webhook-url")
-    SLACK_HEADERS = {
-        'Content-Type': 'application/json'
-    }
-
-    if (
-        event['event_type'] in ['update_progress', 'flow_progress', 'operation_progress']
-        and event['details'].get(event['event_type'], {}).get('state') in ['FAILED', 'STOPPED']
-    ):
-        event_type = event['event_type']
-        state = event['details'].get(event['event_type'], {}).get('state')
-        pipeline_id = event['origin'].get('pipeline_id')
-        pipeline_name = event['origin'].get('pipeline_name')
-        update_id = event['origin'].get('update_id')
-        
-        color = "#FF0000" if state == 'FAILED' else "#FFA500"
-        
-        try:
-            databricks_host = dbutils.secrets.get(scope=f"event-hooks-{ENVIRONMENT}", key="databricks-host")
-        except Exception:
-            databricks_host = None
-        
-        if databricks_host:
-            workspace_url = databricks_host
-            pipeline_url = f"{databricks_host}/pipelines/{pipeline_id}"
-            update_url = f"{databricks_host}/pipelines/{pipeline_id}/updates/{update_id}"
-        else:
-            workspace_url = None
-            pipeline_url = None
-            update_url = None
-        
-        timestamp = event.get('timestamp')
-        if timestamp and isinstance(timestamp, (int, float)):
-            timestamp = datetime.fromtimestamp(timestamp).strftime('%Y-%m-%d %H:%M:%S UTC')
-        
-        def slack_link(url, text):
-            return f"<{url}|{text}>" if url else text
-        
-        payload = {
-            "channel": MONITORING_SLACK_CHANNEL,
-            "text": f"{pipeline_name} — a run has {state.lower()}.",
-            "attachments": [
-                {
-                    "color": color,
-                    "fields": [
-                        {
-                            "title": "Workspace:",
-                            "value": slack_link(workspace_url, f"open-jii-databricks-workspace-{ENVIRONMENT}"),
-                            "short": True
-                        },
-                        {
-                            "title": "Job:",
-                            "value": slack_link(pipeline_url, pipeline_name),
-                            "short": True
-                        },
-                        {
-                            "title": "Update:",
-                            "value": slack_link(update_url, update_id),
-                            "short": True
-                        },
-                        {
-                            "title": "State:",
-                            "value": state,
-                            "short": True
-                        },
-                        {
-                            "title": "Environment:",
-                            "value": ENVIRONMENT.upper(),
-                            "short": True
-                        },
-                        {
-                            "title": "Event Type:",
-                            "value": event_type,
-                            "short": True
-                        }
-                    ],
-                    "actions": [
-                        {
-                            "type": "button",
-                            "text": "View Pipeline",
-                            "url": pipeline_url
-                        },
-                        {
-                            "type": "button", 
-                            "text": "View Workspace",
-                            "url": workspace_url
-                        }
-                    ]
-                }
-            ]
-        }
-        
-        try:
-            response = requests.post(
-                url=SLACK_WEBHOOK_URL,
-                headers=SLACK_HEADERS,
-                json=payload
-            )
-            print(f"Slack notification sent: {event_type} - {state}")
-        except Exception as e:
-            print(f"Failed to send Slack notification: {e}")
